+++ packages/blogpress/.vitepress/theme/index.ts
@@ -29,6 +29,42 @@ export default {
           }
         }
       }
+
+      // 添加博客运行时间计数器
+      var targetDate = new Date('2023-12-29T00:00:00')
+
+      function updateCountdown() {
+        // 当前日期和时间
+        var currentDate = new Date()
+
+        // 计算剩余时间（以毫秒为单位）
+        var timeDiff = currentDate.getTime() - targetDate.getTime()
+
+        // 计算剩余的天、时，分、秒
+        var days = Math.floor(timeDiff / (1000 * 60 * 60 * 24))
+        var hours = Math.floor((timeDiff % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60))
+        var minutes = Math.floor((timeDiff % (1000 * 60 * 60)) / (1000 * 60))
+        var seconds = Math.floor((timeDiff % (1000 * 60)) / 1000)
+
+        // 更新显示剩余时间的容器
+        var blogRunTimeElement = document.getElementById('blogRunTime')
+        if (blogRunTimeElement) {
+          blogRunTimeElement.innerHTML = days + ' d ' + hours + ' h ' + minutes + ' m ' + seconds + ' s '
+        }
+
+        // 获取当前年份
+        var year = currentDate.getFullYear()
+        var yearElement = document.getElementById('year')
+        if (yearElement) {
+          yearElement.innerHTML = String(year)
+        }
+      }
+
+      // 初始调用一次更新函数
+      updateCountdown()
+
+      // 每隔一秒调用一次更新函数
+      setInterval(updateCountdown, 1000)
     }
   }
 }