+++ packages/create-theme/public/template/docs/.vitepress/blog-theme.ts
@@ -73,7 +73,7 @@ const blogTheme = getThemeConfig({
       { type: 'text', content: '👇公众号👇---👇 微信 👇' },
       {
         type: 'image',
-        src: 'https://img.cdn.sugarat.top/mdImg/MTYxNTAxODc2NTIxMA==615018765210~fmt.webp',
+        src: '/zx-vitepress-blog/fmt-code.png',
         alt: '勋染微信二维码'
       },
       {