+++ packages/vitepress-plugin-pagefind/CHANGELOG.md
@@ -292,4 +292,4 @@ The build speed has been significantly improved.
 
 ### Minor Changes
 
-- feat: 从 [@sugarat/theme](https://github.com/ATQQ/sugar-blog/tree/master/packages/theme) 主题分离独立的 pagefind 插件
+- feat: 从 [@sugarat/theme](https://github.com/zhouxun0803/zx-vitepress-blog/tree/master/packages/theme) 主题分离独立的 pagefind 插件